requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
pytesseract>=0.3.10
pyinstaller>=6.0.0
watchdog>=3.0.0
pystray>=0.19.0
